backend/pyproject.toml
[project]
name = "contextual-discord-backend"
version = "0.1.0"
description = "Backend for Contextual Discord Plugin"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sentence-transformers",
    "qdrant-client",
    "requests",
    "python-dotenv",
    "hf_xet",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
